@@ -1,7 +1,7 @@
 package ua.com.integrity.bpm.camunda.study.dto.materail;
 
-import javax.xml.bind.annotation.XmlEnum;
-import javax.xml.bind.annotation.XmlEnumValue;
+import jakarta.xml.bind.annotation.XmlEnum;
+import jakarta.xml.bind.annotation.XmlEnumValue;
 
 @XmlEnum
 public enum MeasureUnit {
