@@ -1,6 +1,6 @@
 package ua.com.integrity.bpm.camunda.study.dto.trip;
 
-import javax.xml.bind.annotation.XmlEnum;
+import jakarta.xml.bind.annotation.XmlEnum;
 
 @XmlEnum
 public enum TripType {
