@@ -14,11 +14,11 @@ import ua.com.integrity.bpm.camunda.study.dto.equipment.SupplyRateDto;
 import ua.com.integrity.bpm.camunda.study.dto.materail.ConsumptionRateDto;
 import ua.com.integrity.bpm.camunda.study.service.EquipmentService;
 
-import javax.inject.Inject;
-import javax.validation.Valid;
-import javax.ws.rs.*;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
+import jakarta.inject.Inject;
+import jakarta.validation.Valid;
+import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import java.util.List;
