@@ -3,10 +3,10 @@ package ua.com.integrity.bpm.camunda.study.api.rest;
 import ua.com.integrity.bpm.camunda.study.dto.trip.TripDto;
 import ua.com.integrity.bpm.camunda.study.service.TripService;
 
-import javax.inject.Inject;
-import javax.validation.Valid;
-import javax.ws.rs.*;
-import javax.ws.rs.core.MediaType;
+import jakarta.inject.Inject;
+import jakarta.validation.Valid;
+import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.MediaType;
 import java.util.List;
 
 @Path("trips")
