@@ -3,11 +3,11 @@ package ua.com.integrity.bpm.camunda.study.api.rest.config;
 import org.apache.commons.lang3.exception.ExceptionUtils;
 import ua.com.integrity.bpm.camunda.study.dto.api.ApiErrorResponse;
 
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.ext.ExceptionMapper;
-import javax.ws.rs.ext.Provider;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.ext.ExceptionMapper;
+import jakarta.ws.rs.ext.Provider;
 
 @Provider
 @Produces(MediaType.APPLICATION_JSON)
