@@ -5,9 +5,9 @@ import ua.com.integrity.bpm.camunda.study.dao.GenericDao;
 import ua.com.integrity.bpm.camunda.study.domain.Employee;
 import ua.com.integrity.bpm.camunda.study.dto.EmployeeDto;
 
-import javax.persistence.EntityManager;
-import javax.persistence.NoResultException;
-import javax.persistence.PersistenceContext;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.NoResultException;
+import jakarta.persistence.PersistenceContext;
 import java.util.List;
 import java.util.Optional;
 
