@@ -6,7 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import ua.com.integrity.bpm.camunda.study.domain.equipment.SupplyRate;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 import java.util.List;
 
