@@ -6,7 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import ua.com.integrity.bpm.camunda.study.domain.materail.ConsumptionRate;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 import java.util.Set;
 
