@@ -13,8 +13,8 @@ import ua.com.integrity.bpm.camunda.study.dto.materail.MaterialDto;
 import ua.com.integrity.bpm.camunda.study.mapper.ErpMapper;
 import ua.com.integrity.bpm.camunda.study.service.MaterialsService;
 
-import javax.inject.Inject;
-import javax.transaction.Transactional;
+import jakarta.inject.Inject;
+import jakarta.transaction.Transactional;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
