@@ -15,8 +15,8 @@ import ua.com.integrity.bpm.camunda.study.dto.vacation.VacationDto;
 import ua.com.integrity.bpm.camunda.study.mapper.ErpMapper;
 import ua.com.integrity.bpm.camunda.study.service.OrgStructureService;
 
-import javax.inject.Inject;
-import javax.transaction.Transactional;
+import jakarta.inject.Inject;
+import jakarta.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
 
